longs

+++ flamebot.py
@@ -42,7 +42,7 @@ class RNN(nn.Module):
         attn_weights = F.softmax(self.attn(torch.cat((embeddings, l1, l2), dim=2)),
                                  dim=0)
         output = self.softmax(self.output(attn_weights))
-        return torch.argmax(output, dim=1)
+        return torch.argmax(output, dim=1).double()
 
 
     def initHidden(self):
